[DDW-60] Cardano Node config cleanup
